Reorganize handlers code.
